LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-006-expected.txt


FAIL #target1 is scroll-adjusted in x axis only assert_equals: expected 150 but got 100
FAIL #target2 is scroll-adjusted in y axis only assert_equals: expected 150 but got 100
PASS #target3 is scroll-adjusted in neither axis
